Rename route files for different actions.
